expect return value instead of mock call, remove duplicate common name mapper
